updated loose ends to work with the included copy of asio in cinder
