+++ examples/exampleTcpApp/vc2019/exampleTcpApp.sln
@@ -5,9 +5,12 @@ MinimumVisualStudioVersion = 10.0.40219.1
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "exampleTcpApp", "exampleTcpApp.vcxproj", "{494DCD97-2AE6-46BF-A05F-6190F32DDC78}"
 	ProjectSection(ProjectDependencies) = postProject
 		{92B5BE70-DCAA-40E4-92D8-CC2B95AA28BE} = {92B5BE70-DCAA-40E4-92D8-CC2B95AA28BE}
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22} = {04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}
 	EndProjectSection
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "cinder", "..\..\..\..\cinder\proj\vc2019\cinder.vcxproj", "{92B5BE70-DCAA-40E4-92D8-CC2B95AA28BE}"
+EndProject
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "sitara-asio", "..\..\..\proj\sitara-asio.vcxproj", "{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -71,6 +74,30 @@ Global
 		{92B5BE70-DCAA-40E4-92D8-CC2B95AA28BE}.Release|x64.Build.0 = Release|x64
 		{92B5BE70-DCAA-40E4-92D8-CC2B95AA28BE}.Release|x86.ActiveCfg = Release|Win32
 		{92B5BE70-DCAA-40E4-92D8-CC2B95AA28BE}.Release|x86.Build.0 = Release|Win32
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Debug_ANGLE|x64.ActiveCfg = Debug|x64
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Debug_ANGLE|x64.Build.0 = Debug|x64
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Debug_ANGLE|x86.ActiveCfg = Debug|Win32
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Debug_ANGLE|x86.Build.0 = Debug|Win32
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Debug_Shared|x64.ActiveCfg = Debug|x64
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Debug_Shared|x64.Build.0 = Debug|x64
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Debug_Shared|x86.ActiveCfg = Debug|Win32
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Debug_Shared|x86.Build.0 = Debug|Win32
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Debug|x64.ActiveCfg = Debug|x64
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Debug|x64.Build.0 = Debug|x64
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Debug|x86.ActiveCfg = Debug|Win32
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Debug|x86.Build.0 = Debug|Win32
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Release_ANGLE|x64.ActiveCfg = Release|x64
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Release_ANGLE|x64.Build.0 = Release|x64
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Release_ANGLE|x86.ActiveCfg = Release|Win32
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Release_ANGLE|x86.Build.0 = Release|Win32
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Release_Shared|x64.ActiveCfg = Release|x64
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Release_Shared|x64.Build.0 = Release|x64
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Release_Shared|x86.ActiveCfg = Release|Win32
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Release_Shared|x86.Build.0 = Release|Win32
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Release|x64.ActiveCfg = Release|x64
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Release|x64.Build.0 = Release|x64
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Release|x86.ActiveCfg = Release|Win32
+		{04DB14E6-1ECC-4B8C-AA98-00BCFAF03E22}.Release|x86.Build.0 = Release|Win32
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
