+++ proj/sitara-asio.cpp
@@ -1,0 +1,9 @@
+// sitara-asio.cpp : Defines the functions for the static library.
+//
+
+#include "framework.h"
+
+// TODO: This is an example of a library function
+void fnsitaraasio()
+{
+}
