+++ src/udp/UdpSocket.cpp
@@ -1,4 +1,4 @@
-#include "UdpSocket.h"
+#include "udp/UdpSocket.h"
 #include <iostream>
 
 using namespace sitara;
